Added tenant service interface in startup to resolve some errors

+++ BeanBag/Services/IPaymentService.cs
@@ -10,7 +10,7 @@ namespace BeanBag.Services
         Dictionary<string, string> ToDictionary(string response);
         string GetMd5Hash(Dictionary<string, string> data, string encryptionKey);
         bool VerifyMd5Hash(Dictionary<string, string> data, string encryptionKey, string hash);
-        public bool AddTransaction(string reference, string payId, string tenantId, double amount);
+        public bool AddTransaction(string reference, string payId, string tenantId, float amount);
         IEnumerable<Transaction> GetTransactions(string currentTenantId);
     }
 }
